Latitud y Longitud a Double

+++ sportex.backoffice/Models/LocationDTO.cs
@@ -11,8 +11,8 @@ namespace sportex.backoffice.Models
         public int ID { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
-        public int? Latitude { get; set; }
-        public int? Longitude { get; set; }
+        public double? Latitude { get; set; }
+        public double? Longitude { get; set; }
         #endregion
 
         public LocationDTO()
